[FEAT] 목표 달성 여부 조회하는 모든 api에 배치 없이 돌아가는 로직 적용
