+++ src/controller/goalController.ts
@@ -233,6 +233,8 @@ const getHistoryByGoalId = async (req: Request, res: Response) => {
 }
 
 const achieveGoal = async (req: Request, res: Response) => {
+  const now = date.getNow();
+  const nowPlus9h = date.getCurrentDatePlus9h(now);
   const userId = req.user.userId;
   const { goalId } = req.params;
   const { isAchieved } = req.body;
@@ -245,7 +247,7 @@ const achieveGoal = async (req: Request, res: Response) => {
 
   let data;
   try {
-    data = await goalService.achieveGoal(+goalId, isAchieved as boolean);
+    data = await goalService.achieveGoal(+goalId, isAchieved as boolean, now, nowPlus9h);
 
     if (data === goalError.DOUBLE_ACHIEVED_ERROR) {
       slack.sendErrorMessageToSlack(
@@ -298,6 +300,7 @@ const sortType = {
 };
 
 const getKeptGoalsByUserId = async (req: Request, res: Response) => {  
+  const now = date.getNow();
   const userId = req.user.userId;
   const sort = req.query.sort as string;
 
@@ -315,7 +318,7 @@ const getKeptGoalsByUserId = async (req: Request, res: Response) => {
 
   try {
   
-    const foundGoals = await goalService.getKeptGoals(+userId, sort as string);
+    const foundGoals = await goalService.getKeptGoals(+userId, sort as string, now);
     
     return res
       .status(sc.OK)
